Run button for `SPARQL` example queries (#12)
There is now a run button for the SPARQL example queries. It is placed on the top right of the code block panel, just to the left of the copy-and-paste icon. Link to example video below. Here is an example for how this is done in the source. Closes #11
````markdown
```sparql {data-demo-engine="osm-planet"}
SELECT * WHERE {
  ?s ?p ?o
} LIMIT 10
```
````
https://github.com/user-attachments/assets/7d0a6bb1-600c-47ae-bf59-5965d94cf3fe